Restructure files and add nav bar to resetPassword (#6)
